% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hummus_objet.R
\name{add_network}
\alias{add_network}
\title{Add a network to a multiplex, a multilayer or an hummus object}
\usage{
add_network(
  object,
  network,
  network_name,
  multiplex_name = NULL,
  directed = FALSE,
  weighted = FALSE,
  verbose = 1
)
}
\arguments{
\item{object}{A multiplex, a multilayer or an hummus object}

\item{network}{A network (edge list)}

\item{network_name}{The name of the network}

\item{multiplex_name}{The name of the multiplex. Default: NULL if object is a
multiplex already only}

\item{directed}{Logical indicating if the network is directed. Default: FALSE}

\item{weighted}{Logical indicating if the network is weighted. Default: FALSE}

\item{verbose}{(integer) - Display function messages. Set to 0 for no
message displayed, >= 1 for more details.}
}
\value{
A multiplex, a multilayer or an hummus object with the added network
}
\description{
Add a network to a multiplex, a multilayer or an hummus object
}
\examples{
hummus <- add_network(
                           object = hummus,
                           network = atac_peak_network,
                           network_name = network_name,
                           multiplex_name = multiplex_name,
                           weighted = TRUE,
                           directed = FALSE)

}
