% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hummus_objet.R
\docType{class}
\name{bipartite-class}
\alias{bipartite-class}
\alias{bipartite}
\title{Bipartite class}
\description{
Bipartite object stores a bipartite network (edge list) and the names of the
left and right features' multiplexes.
}
\section{Slots}{

\describe{
\item{\code{network}}{(data.frame) - Bipartite network (edge list)}

\item{\code{multiplex_left}}{(character) - Left features' multiplex}

\item{\code{multiplex_right}}{(character) - Right features' multiplex}
}}

\examples{
bipartite <- bipartite(
                          network = bipartite_network,
                         multiplex_left = "RNA",
                        multiplex_right = "peaks")

}
