% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hummus_objet.R
\name{DefaultAssay.Hummus_Object}
\alias{DefaultAssay.Hummus_Object}
\title{Get Default assays of Hummus_Object (based on Seurat)}
\usage{
\method{DefaultAssay}{Hummus_Object}(object, ...)
}
\description{
Get Default assays of Hummus_Object (based on Seurat)
}
\examples{
# Get current default assay
DefaultAssay(object = pbmc_small)

}
