% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_network.R
\name{define_general_config}
\alias{define_general_config}
\title{Define general config file for hummuspy}
\usage{
define_general_config(
  hummus_object,
  multiplex_names = NULL,
  bipartites_names = NULL,
  folder_multiplexes = "multiplex",
  folder_bipartites = "bipartites",
  seed_path = "seed/seeds.txt",
  suffix = ".tsv",
  self_loops = FALSE,
  restart_proba = 0.7,
  save_configfile = FALSE,
  config_name = "config.yml",
  config_folder = "config",
  multilayer_f = "multilayer"
)
}
\arguments{
\item{hummus_object}{A hummus object}

\item{multiplex_names}{A vector of multiplex names considered.
It must be a subset of the names of the multiplexes in the hummus object, or NULL
if all multiplexes should be considered.}

\item{bipartites_names}{A vector of bipartites names considered.
It must be a subset of the names of the bipartites in the hummus object, or NULL
if all bipartites should be considered.}

\item{folder_multiplexes}{The folder where the multiplexes are stored}

\item{folder_bipartites}{The folder where the bipartites are stored}

\item{seed_path}{The path to the seed file}

\item{self_loops}{A boolean indicating if self loops should be considered.}

\item{restart_proba}{The restart probability for the random walk (default = 0.7)}

\item{save_configfile}{A boolean indicating if the config file should be saved}

\item{config_name}{The name of the config file to be created by hummuspy}

\item{config_folder}{The folder where the config file will be created (inside multilayer_f)}

\item{multilayer_f}{The folder where the multilayer is stored}

\item{suffix_bipartites}{A suffix to add to the bipartites names (to indicate
the exact file name)}
}
\value{
A config file for hummuspy
}
\description{
Define general config file for hummuspy
}
