% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_network.R
\name{define_grn}
\alias{define_grn}
\title{Define GRN from hummus object}
\usage{
define_grn(
  hummus_object,
  multiplex_names = NULL,
  bipartites_names = NULL,
  config_name = "grn_config.yml",
  config_folder = "config",
  tf_multiplex = "TF",
  atac_multiplex = "peaks",
  rna_multiplex = "RNA",
  multilayer_folder = "multilayer",
  gene_list = NULL,
  tf_list = NULL,
  save = FALSE,
  output_f = NULL,
  return_df = TRUE,
  suffix_bipartites = ".tsv",
  njobs = 1
)
}
\arguments{
\item{hummus_object}{A hummus object}

\item{multiplex_names}{A vector of multiplex names considered.
It must be a subset of the names of the multiplexes in the hummus object.}

\item{bipartites_names}{A vector of bipartites names considered.
It must be a subset of the names of the bipartites in the hummus object.}

\item{config_name}{The name of the config file to be created by hummuspy}

\item{config_folder}{The folder where the config file will be created}

\item{tf_multiplex}{The name of the multiplex containing the TFs}

\item{atac_multiplex}{The name of the multiplex containing the ATAC-seq peaks}

\item{rna_multiplex}{The name of the multiplex containing the RNA-seq genes}

\item{multilayer_folder}{The path to the folder where the multilayer processed files are stored}

\item{gene_list}{A vector of genes to be considered for the final GRN
(filtering is done on the genes before inferring the GRN)}

\item{tf_list}{A vector of TFs to be considered for the final GRN (filtering
is done on the TFs after inferring the GRN)}

\item{save}{A boolean indicating if the GRN should be saved}

\item{output_f}{The name of the file where the GRN should be saved
(if save == TRUE)}

\item{return_df}{A boolean indicating if the GRN should be returned as a
dataframe}

\item{suffix_bipartites}{A suffix to add to the bipartites names (to indicate
the exact file location)}

\item{njobs}{The number of jobs to be used for the computation of the GRN}
}
\value{
A dataframe containing the GRN (if return_df == TRUE)
}
\description{
Calling the define_output function with output_type = 'GRN'
}
\examples{
grn <- define_grn(hummus_object = hummus,
                            multilayer_folder = multilayer_folder,
                            njobs = 5)

}
