% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{extract_names}
\alias{extract_names}
\title{Extract multilayer and bipartites from hummus object}
\usage{
extract_names(hummus_object, layers = NULL, bipartites = NULL)
}
\description{
Extract multilayer and bipartites from hummus object,
in order to use them in the python script building the config file.

param hummus_object: hummus object
param layers: list of layers to extract if not all
param bipartites: list of bipartites to extract if not all
}
