% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_peaks_near_genes}
\alias{find_peaks_near_genes}
\title{Title}
\usage{
find_peaks_near_genes(
  peaks,
  genes,
  sep = c("-", "-"),
  method = c("Signac", "GREAT"),
  upstream = 1e+05,
  downstream = 0,
  extend = 1e+06,
  only_tss = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{peaks}{TODO}

\item{genes}{TODO}

\item{sep}{TODO}

\item{method}{TODO}

\item{upstream}{TODO}

\item{downstream}{TODO}

\item{extend}{TODO}

\item{only_tss}{TODO}

\item{verbose}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
