% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_network.R
\name{format_bipartites_names}
\alias{format_bipartites_names}
\title{Format bipartites names for python hummuspy package config functions}
\usage{
format_bipartites_names(
  hummus_object,
  bipartites_names = NULL,
  suffix_bipartites = ".tsv"
)
}
\arguments{
\item{hummus_object}{A hummus object}

\item{bipartites_names}{A vector of bipartites names considered.
It must be a subset of the names of the bipartites in the hummus object.}

\item{suffix_bipartites}{A suffix to add to the bipartites location}
}
\value{
A list of bipartites names formatted for hummuspy config funtions
each element of the list is a list containing the right and left layer
connected by the bipartite
}
\description{
Format bipartites names for python hummuspy package config functions
}
\examples{
bipartites_dictionary <- format_bipartites_names(
                                      hummus_object = hummus,
                                      bipartites_names = c("atac_rna",
                                                           "tf_peaks"))

}
