% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_network.R
\name{format_multiplex_names}
\alias{format_multiplex_names}
\title{Format multiplex names for python hummuspy package config functions}
\usage{
format_multiplex_names(hummus_object, multiplex_names = NULL)
}
\arguments{
\item{hummus_object}{A hummus object}

\item{multiplex_names}{A vector of multiplex names considered. It must be
a subset of the names of the multiplexes in the hummus object.}
}
\value{
A list of multiplexes names formatted for hummuspy config funtions
each element of the list is a list of the network types (weighted/directed)
and the name of the networks as named in the hummus object
}
\description{
Format multiplex names for python hummuspy package config functions
}
\examples{
multiplexes_dictionary <- format_multiplex_names(
                                       hummus_object = hummus,
                                       multiplex_names = c("TF", "peaks"))

}
