% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_online.R
\name{get_genome_annotations}
\alias{get_genome_annotations}
\title{Fetch online genome annotations from Ensembldb database}
\usage{
get_genome_annotations(
  ensdb_annotations = EnsDb.Hsapiens.v86::EnsDb.Hsapiens.v86
)
}
\arguments{
\item{EnsDb_annotations}{(EndsDb object) - Ensembldb database (default: EnsDb.Hsapiens.v86::EnsDb.Hsapiens.v86}
}
\value{
gene_range (GRanges object) - Genome annotations
}
\description{
Fetch online genome annotations from Ensembldb database
}
\examples{
gene_range = get_genome_annotations(EnsDb.Hsapiens.v86::EnsDb.Hsapiens.v86)
}
