% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_online.R
\name{get_tf2motifs}
\alias{get_tf2motifs}
\title{Fetch online TF motifs from JASPAR2020 and chromVARmotifs}
\usage{
get_tf2motifs(species = "human")
}
\arguments{
\item{species}{TODO}
}
\value{
TODO
}
\description{
Fetch online TF motifs from JASPAR2020 and chromVARmotifs
}
\examples{
TODO
}
