% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTFs}
\alias{getTFs}
\title{Title}
\usage{
getTFs(
  species = "human",
  l_expressed_genes,
  output_file,
  tf2motifs,
  verbose = 1
)
}
\arguments{
\item{species}{TODO}

\item{l_expressed_genes}{TODO}

\item{output_file}{TODO}

\item{tf2motifs}{TODO}

\item{verbose}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
