% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hummus_objet.R
\name{Initiate_Hummus_Object}
\alias{Initiate_Hummus_Object}
\title{Initiate a hummus object}
\usage{
Initiate_Hummus_Object(
  seurat_assays,
  active.assay = NULL,
  multilayer = NULL,
  motifs_db = NULL
)
}
\arguments{
\item{seurat_assays}{A Seurat object or a list of Seurat assays}

\item{active.assay}{The name of the active assay. Default: NULL}

\item{multilayer}{A multilayer object. Default: NULL}

\item{motifs_db}{A motifs_db object. Default: NULL}
}
\value{
A hummus object
}
\description{
Initiate a hummus object
}
\examples{
seurat_object <- Seurat::CreateSeuratObject(counts = matrix(rnorm(1000), nrow = 100, ncol = 10))
         hummus <- InitiateHummus_Object(seurat_object)
         hummus
}
