% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{class}
\name{MotifsDatabase-class}
\alias{MotifsDatabase-class}
\alias{MotifsDatabase}
\title{The Modules class}
\description{
The Modules object stores the TF modules extracted from the inferred network..
}
\section{Slots}{

\describe{
\item{\code{motifs}}{PWMatrixList.}

\item{\code{tf2motifs}}{data.frame.}
}}

