% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hummus_objet.R
\docType{class}
\name{multiplex-class}
\alias{multiplex-class}
\alias{multiplex}
\title{Multiplex class}
\description{
Multiplex object stores a list of networks, a list of features and
a list of logicals indicating if the network is directed or weighted.
}
\section{Slots}{

\describe{
\item{\code{networks}}{(list) - List of networks.}

\item{\code{features}}{(vector) - Vector of features.}

\item{\code{directed}}{(list) - List of logical indicating if networks are directed.}

\item{\code{weighted}}{(list) - List of logical indicating if networks are weighted.}
}}

