% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{peaks_in_regulatory_elements}
\alias{peaks_in_regulatory_elements}
\title{Filter peaks to those overlapping specific (regulatory) elements}
\usage{
peaks_in_regulatory_elements(
  Peaks,
  RegEl,
  sep_Peak1 = "-",
  sep_Peak2 = "-",
  sep_RegEl1 = "-",
  sep_RegEl2 = "-"
)
}
\arguments{
\item{Peaks}{(character) vector of genomic coordinates of peaks}

\item{RegEl}{(character) vector of genomic coordinates of regulatory elements}

\item{sep_Peak1}{(character) separator between chromosome and
start position of peak}

\item{sep_Peak2}{(character) separator between start position
and end position of peak}

\item{sep_RegEl1}{(character) separator between chromosome and
start position of regulatory element}

\item{sep_RegEl2}{(character) separator between start position and
end position of regulatory element}
}
\value{
(character) vector of genomic coordinates of peaks overlapping
}
\description{
Function to reduce list of "Peaks" to the ones overlapping with
list of "RegEl", e.g. regulatory elements, evolutionary conserved regions
}
\examples{
peaks_in_regulatory_elements(peaks, RegEl)
}
