% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{peaksInRegEl}
\alias{peaksInRegEl}
\title{Filter peaks to those overlapping specific element}
\usage{
peaksInRegEl(
  Peaks,
  RegEl,
  sep_Peak1 = "-",
  sep_Peak2 = "-",
  sep_RegEl1 = "-",
  sep_RegEl2 = "-"
)
}
\arguments{
\item{Peaks}{TODO}

\item{RegEl}{TODO}

\item{sep_Peak1}{TODO}

\item{sep_Peak2}{TODO}

\item{sep_RegEl1}{TODO}

\item{sep_RegEl2}{TODO}
}
\value{
TODO
}
\description{
function to reduce list of "Peaks" to the ones overlapping with list of "RegEl"
(e.g. regulatory elements, evolutionary conserved regions)
}
\examples{
TODO
}
