% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_wrappers.R
\name{run_tf_null_wrapper}
\alias{run_tf_null_wrapper}
\title{tf_null wrapper function for the tf_network function}
\usage{
run_tf_null_wrapper(
  hummus = hummus,
  organism = organism,
  tfs = tfs,
  gene_assay = gene_assay,
  verbose = 1
)
}
\arguments{
\item{hummus}{A hummus object}
}
\description{
This function is a wrapper for the tf_network function
}
