% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hummus_objet.R
\name{save_multilayer}
\alias{save_multilayer}
\title{Save multilayer object files in a hierarchical structure on disk}
\usage{
save_multilayer(hummus, folder_name, verbose = TRUE, suffix = ".tsv")
}
\arguments{
\item{hummus}{A hummus object}

\item{folder_name}{The name of the folder to save the multilayer}

\item{verbose}{(integer) - Display function messages. Set to 0 for no
message displayed, >= 1 for more details.}

\item{suffix}{The suffix of the files to save. Default: ".tsv"}
}
\value{
Nothing, but create a folder containing the multilayer object files
}
\description{
Save multilayer files from a hummus_object
in a hierarchical structure on disk, inside a folder specified through
folder_name
}
\examples{
folder_name = "multilayer"
save_multilayer(hummus = hummus, folder_name = "multilayer")

}
