% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hummus_objet.R
\name{store_network}
\alias{store_network}
\title{Wrapper function to save a network or not}
\usage{
store_network(network, store_network, output_file, verbose = 1)
}
\arguments{
\item{network}{A network (edge list)}

\item{store_network}{Logical indicating if the network should be saved}

\item{output_file}{The name of the file to save the network}

\item{verbose}{(integer) - Display function messages. Set to 0 for no
message displayed, >= 1 for more details.}
}
\value{
Nothing, but save the network in a file if store_network is TRUE
}
\description{
Wrapper function to save a network or not in a file according to the
store_network parameter. If store_network is TRUE, the network is saved in
the output_file.
}
\examples{
network <- read.table("network.tsv", header = TRUE, sep = "\t")
          store_network(network = network,
              store_network = TRUE,
              output_file = "network.tsv",
              verbose = 1)

}
