% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hummus_objet.R
\name{[[<-,Hummus_Object}
\alias{[[<-,Hummus_Object}
\alias{[[.Hummus_Object}
\alias{[[<-.Hummus_Object}
\alias{\S4method{[[<-}{Hummus_Object,character,missing,Assay}}
\title{Access assays in Hummus_Object (based on Seurat)}
\usage{
\method{[[}{Hummus_Object}(x, i = missing_arg(), ..., drop = FALSE, na.rm = FALSE)
}
\description{
Access assays in Hummus_Object (based on Seurat)
}
