% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hummus_objet.R
\name{[[<-,Seurat}
\alias{[[<-,Seurat}
\alias{VariableFeatures.Hummus_Object}
\alias{[[<-.Hummus_Object}
\alias{\S4method{[[<-}{Hummus_Object,character,missing,Assay}}
\title{Variable features of assays in Hummus_Object (based on Seurat)}
\usage{
\method{VariableFeatures}{Hummus_Object}(
  object,
  method = NULL,
  assay = NULL,
  nfeatures = NULL,
  layer = NA,
  simplify = TRUE,
  selection.method = lifecycle::deprecated(),
  ...
)
}
\description{
Variable features of assays in Hummus_Object (based on Seurat)
}
