% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{compute_atac_peak_network}
\alias{compute_atac_peak_network}
\title{Compute peak network from scATAC-seq data}
\usage{
compute_atac_peak_network(
  scATAC,
  genome = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38,
  method = "cicero",
  store_network = TRUE,
  output_file,
  threshold = 0,
  number_cells_per_clusters = 50,
  sample_num = 100,
  seed = 2021,
  verbose = 1,
  window = 5e+05
)
}
\arguments{
\item{scATAC}{TODO}

\item{genome}{TODO}

\item{method}{TODO}

\item{store_network}{(bool) - Save the network directly (\code{TRUE},
default) or return without saving on disk (\code{FALSE}).}

\item{output_file}{(character) - Name of the output_file (if store_network == \code{TRUE}).}

\item{threshold}{(interger, default 0) - Minimal threshold to select tf-gene edges.}

\item{number_cells_per_clusters}{(integer) - Number of cells grouped by territory to define pseudocells}

\item{sample_num}{(integer | Cicero) - TO DO.}

\item{seed}{( __ ) - Random seed used by Cicero.}

\item{verbose}{(integer) - Display function messages. Set to 0 for no message displayed, >= 1 for more details.}

\item{window}{(interger) - Size of window to consider potential cis-regulatory cooperations}
}
\value{
(data.frame) - Return list of network interactions between peaks
}
\description{
This function will create a network from atac data (or in theory any data
wtih peaks coordinates as features).
Different method should be implemented at some point (e.g. RENIN),
for now Cicero is still the reference and only method available
}
\details{
Method descriptions :
\enumerate{
\item Cicero
Use patial corelation between peaks that are in a given window (e.g. :
less distant than 500K base pairs)
}
}
\examples{
TO DO. Same than UNIT test.
}
