% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{compute_gene_network}
\alias{compute_gene_network}
\title{Compute gene netwok from scRNA-seq data}
\usage{
compute_gene_network(
  hummus,
  gene_assay = "RNA",
  tfs = NULL,
  method = "GENIE3",
  store_network = FALSE,
  output_file,
  threshold = 0,
  number_cores = 1,
  verbose = 1,
  multiplex_name = NULL
)
}
\arguments{
\item{tfs}{vector(character) - List of tfs considered.}

\item{method}{(character) - Method used to infer network edges.
\itemize{
\item \code{'Genie3'} - TO DO.
}}

\item{store_network}{(bool) - Save the network directly (\code{TRUE},
default) or return without saving on disk (\code{FALSE}).}

\item{output_file}{(character) - Name of the output_file (if store_network == \code{TRUE}).}

\item{threshold}{(interger, default 0) - Minimal threshold to select tf-gene edges.}

\item{number_cores}{(interger, default 1) - Number of thread that should be used for the parallelizable methods.}

\item{verbose}{(integer) - Display function messages. Set to 0 for no message displayed, >= 1 for more details.}

\item{scRNA}{(data.frame) - Expression matrix (cells*genes)}
}
\value{
(data.frame) - Return list of network interactions between genes
}
\description{
This function will create a network from rna data (or in theory any data
wtih genes as features).
Different method should be implemented at some point (any suggestion is welcomed ! :) ),
for now Genie3 is still the reference and only method available
}
\details{
Method descriptions :
\enumerate{
\item Genie3
Use tree random forest to infer regulatory networks :
https://bioconductor.org/packages/release/bioc/html/GENIE3.html
}
}
\examples{
TO DO. Same than UNIT test.
}
