% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hummus_objet.R
\name{DefaultAssay}
\alias{DefaultAssay}
\alias{DefaultAssay.Hummus_Object}
\alias{DefaultAssay<-}
\title{Get Default assays of Hummus_Object (based on Seurat)}
\usage{
\method{DefaultAssay}{Hummus_Object}(object, ...)

DefaultAssay(object, ...)

DefaultAssay(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{value}{Name of assay to set as default}
}
\value{
\code{DefaultAssay}: The name of the default assay

\code{DefaultAssay<-}: An object with the default assay updated
}
\description{
Get and set the default assay
}
\examples{
# Get current default assay
DefaultAssay(object = pbmc_small)

}
\concept{data-access}
