% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_peaks_near_genes}
\alias{find_peaks_near_genes}
\title{Associate peaks to genes based on distance to TSS (or gene body)}
\usage{
find_peaks_near_genes(
  peaks,
  genes,
  sep = c("-", "-"),
  method = c("Signac", "GREAT"),
  upstream = 1e+05,
  downstream = 0,
  extend = 1e+06,
  only_tss = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{peaks}{vector(character) - List of peaks.}

\item{genes}{vector(character) - List of genes.}

\item{sep}{vector(character) - Separator between chromosome,
start and end position. Default: c('-', '-').}

\item{method}{(character) - Method to use. Default: "Signac".
\itemize{
\item \code{'Signac'} - Use Signac::Extend to extend genes.
\item \code{'GREAT'} - Not implemented yet.
}}

\item{upstream}{(int) - Upstream distance from TSS
to consider as potential promoter.}

\item{downstream}{(int) - Downstream distance from TSS
to consider as potential promoter.}

\item{extend}{(int) - Integer defining the distance from the upstream
and downstream of the basal regulatory region. Used only by method 'GREAT'.}

\item{only_tss}{(logical) - If TRUE, only TSS will be considered.}

\item{verbose}{(logical) - If TRUE, print progress messages.}
}
\value{
(matrix) - Matrix of peaks x genes with 1 if peak is near gene.
}
\description{
Associate peaks to genes based on distance to TSS (or gene body)
}
\examples{
TODO
}
