% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_online.R
\name{get_genome_annotations}
\alias{get_genome_annotations}
\title{Title}
\usage{
get_genome_annotations(EnsDb_annotations = EnsDb.Hsapiens.v86)
}
\arguments{
\item{EnsDb_annotations}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
