% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTFs}
\alias{getTFs}
\title{Extract TF names from scRNA data and tf2motifs}
\usage{
getTFs(
  species = "human",
  l_expressed_genes,
  output_file,
  tf2motifs,
  verbose = 1
)
}
\arguments{
\item{species}{(character) - Species name. Default: "human".}

\item{l_expressed_genes}{(vector(character)) - List of expressed genes.}

\item{output_file}{(character) - Path to output file.}

\item{tf2motifs}{(data.frame) - TF to motifs names mapping. Columns: motif, tf.}

\item{verbose}{(integer) - Verbosity level. Default: 1.}
}
\value{
TFs (vector(character)) - List of TFs expressed with motifs in the database.
}
\description{
Extract TF names from scRNA data and tf2motifs
}
