% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hummus_objet.R
\docType{class}
\name{hummus_object-class}
\alias{hummus_object-class}
\alias{hummus_object}
\title{The hummus_object class}
\description{
The hummus_object object is an extended \code{Seurat} object
for the storage and analysis of a heterogeneous multilayer network
}
\section{Slots}{

\describe{
\item{\code{multilayer}}{(multilayer) - Multilayer object}

\item{\code{motifs_db}}{(motifs_db) - Motifs database}

\item{\code{assay}}{(list) - List of assays}
}}

\examples{
hummus_object <- hummus_object(seurat_object)

}
