% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hummus_objet.R
\docType{class}
\name{Hummus_Object-class}
\alias{Hummus_Object-class}
\alias{Hummus_Object}
\title{The Hummus_Object class}
\description{
The Hummus_Object object is an extended \code{Seurat} object
for the storage and analysis of a heterogeneous multilayer network
}
\section{Slots}{

\describe{
\item{\code{multilayer}}{(multilayer) - Multilayer object}

\item{\code{motifs_db}}{(motifs_db) - Motifs database}

\item{\code{assay}}{(list) - List of assays}
}}

