% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hummus_objet.R
\name{VariableFeatures}
\alias{VariableFeatures}
\alias{VariableFeatures.Hummus_Object}
\alias{VariableFeatures<-}
\title{Variable features of assays in Hummus_Object (based on Seurat)}
\usage{
\method{VariableFeatures}{Hummus_Object}(
  object,
  method = NULL,
  assay = NULL,
  nfeatures = NULL,
  layer = NA,
  simplify = TRUE,
  selection.method = lifecycle::deprecated(),
  ...
)

VariableFeatures(object, method = NULL, ...)

VariableFeatures(object, ...) <- value
}
\arguments{
\item{value}{A character vector of variable features}
}
\value{
\code{VariableFeatures}: a vector of the variable features
}
\description{
Variable features of assays in Hummus_Object (based on Seurat)
}
